requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-secret-manager==2.16.3
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
proto-plus==1.22.3
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.4.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.16
